requirements.txt
flask==2.3.2
pandas==2.0.3
pyspark==3.5.0
transformers==4.34.0
snowflake-connector-python==3.0.3
scikit-learn==1.3.1
prometheus-flask-exporter==1.7.1
